backend/requirements.txt
fastapi[all]
tortoise-orm[asyncmy]
